// src/types/LanguageType.ts
export enum LanguageType {
    DK = 'DA',
    GR = 'EL',
    DE = 'DE',
    US = 'EN',
    ES = 'ES',
    ET = 'ET',
    FI = 'FI',
    FR = 'FR',
    IT = 'IT',
    NL = 'NL',
    NO = 'NO',
    NN = 'NN',
    PT = 'PT',
    SE = 'SE',
    SV = 'SV',
    UA = 'UK',
    RU = 'RU',
    AR = 'AR',
    BG = 'BG',
    CS = 'CS',
    HR = 'HR',
    HU = 'HU',
    JA = 'JA',
    KA = 'KA',
    PL = 'PL',
    BR = 'BR',
    RO = 'RO',
    SK = 'SK',
    SL = 'SL',
    TR = 'TR',
    XA = 'XA',
    ZH = 'ZH',
    HK = 'HK',
    LV = 'LV',
}